src/acquila_zmq/pyproject.toml
[project]
name = "acquila-zmq"
version = "1.0.1"
description = "ZMQ communication library for Acquila components"
requires-python = ">=3.11"
dependencies = [
    "pyzmq>=27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["acquila_zmq"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
